+++ envs/.env.stage
@@ -1,0 +1,3 @@
+ENV=DEBUG
+
+CLIENT_URL=http://124.197.210.234:3000/
